refactor: 별도 TemplateSelect로 분리하고 step 추가 (#35)
